requirements.txt
openai
datasets==2.19.0
python-dotenv